Kludge Test_SuperLU_DIST to get around bug #1142 - SuperluDIST failure on dense matrices - on atlantis

+++ packages/amesos/src/Amesos_Dscpack.cpp
@@ -352,7 +352,7 @@ int Amesos_Dscpack::PerformNumericFactorization()
   }
 
   DscMap_ = new Epetra_Map(numrows, NumLocalCols, LocalStructOldNum, 0, Comm());
-  AMESOS_CHK_ERR( ( (int) DscMap_ == 0 ) ) ; 
+  AMESOS_CHK_ERR( ( DscMap_ == 0 ) ) ; 
   
   //
   //  Import from the CrsMatrix
